+++ src/main/java/org/inventors/ftc/robotbase/RobotEx.java
@@ -12,13 +12,13 @@ import com.arcrobotics.ftclib.command.button.Trigger;
 import com.arcrobotics.ftclib.gamepad.GamepadKeys;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.inventors.ftc.robotbase.Driving.DriveConstants;
-import org.inventors.ftc.robotbase.Driving.HeadingControllerSubsystem;
-import org.inventors.ftc.robotbase.Driving.MecanumDriveCommand;
-import org.inventors.ftc.robotbase.Driving.MecanumDrivePPV2;
-import org.inventors.ftc.robotbase.Sensors.Camera;
-import org.inventors.ftc.robotbase.Sensors.IMUSubsystem;
-import org.inventors.ftc.robotbase.util.GamepadExEx;
+import org.inventors.ftc.robotbase.drivebase.DriveConstants;
+import org.inventors.ftc.robotbase.drivebase.HeadingControllerSubsystem;
+import org.inventors.ftc.robotbase.drivebase.MecanumDriveCommand;
+import org.inventors.ftc.robotbase.drivebase.MecanumDrivePPV2;
+import org.inventors.ftc.robotbase.sensors.Camera;
+import org.inventors.ftc.robotbase.sensors.IMUSubsystem;
+import org.inventors.ftc.robotbase.hardware.GamepadExEx;
 import org.inventors.ftc.robotbase.util.TelemetrySubsystem;
 
 import java.util.function.BooleanSupplier;
