+++ src/main/java/org/inventors/ftc/robotbase/hardware/GamepadExEx.java
@@ -1,4 +1,4 @@
-package org.inventors.ftc.robotbase.util;
+package org.inventors.ftc.robotbase.hardware;
 
 import com.arcrobotics.ftclib.gamepad.GamepadEx;
 import com.qualcomm.robotcore.hardware.Gamepad;
